Fixed bug

+++ src/main/java/com/forms/formswebapp/user/domain/UserService.java
@@ -11,7 +11,6 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -58,14 +57,22 @@ public class UserService implements UserDetailsService {
 
     public void deleteUser(final String email) {
         final User user = getUserByEmailOrThrow(email);
+        this.checkIfNotAdmin(user);
         userRepository.delete(user);
     }
 
     public UserDto promoteUser(final String email) {
         final User user = getUserByEmailOrThrow(email);
+        this.checkIfNotAdmin(user);
         user.setRole(Role.ADMIN);
         final User saved = userRepository.save(user);
-        return UserDto.from(user);
+        return UserDto.from(saved);
+    }
+
+    private void checkIfNotAdmin(final User user) {
+        if (user.getRole() == Role.ADMIN) {
+            throw new IllegalStateException("Cannot perform this operation on admin user");
+        }
     }
 
 }
